refactor MenuDrawer into granular components for codebase clarity
